requirements.txt
kafka-python
opensearch-py
PyYAML

